Make BKTree fields private
